checks for potential error

+++ deepRetinotopy.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+exec > >(tee -a deepRetinotopy_output.log)
+exec 2> >(tee -a deepRetinotopy_error.log >&2)
 
 while getopts s:t:d:m: flag
 do
@@ -21,22 +23,28 @@ echo "Maps: ${maps[@]}";
 cd main
 for hemisphere in 'lh' 'rh';
 do 
-    # Inside the container
     echo "Generating mid-thickness surface and curvature data..."
     bash 1_native2fsaverage.sh -s $dirSubs -t $dirHCP -h $hemisphere 
 
-    for map in "${maps[@]}";
-    do
-
-        echo "Retinotopy prediction..."
-        python 2_inference.py --path $dirSubs --dataset $datasetName --prediction_type $map --hemisphere $hemisphere
-
-        rm -r $dirSubs/processed
-
-        echo "Resampling predictions to native space..."
-        for model in model1 model2 model3 model4 model5 average;
+    if [ $? -eq 1 ]; then
+        echo "Error in 1_native2fsaverage.sh"
+        exit 1
+    else
+        for map in "${maps[@]}";
         do
-            bash 3_fsaverage2native.sh -s $dirSubs -t $dirHCP -h $hemisphere -r $map -m $model
+
+            echo "Retinotopy prediction..."
+            python 2_inference.py --path $dirSubs --dataset $datasetName --prediction_type $map --hemisphere $hemisphere
+
+            rm -r $dirSubs/processed
+
+            echo "Resampling predictions to native space..."
+            for model in model1 model2 model3 model4 model5 average;
+            do
+                bash 3_fsaverage2native.sh -s $dirSubs -t $dirHCP -h $hemisphere -r $map -m $model
+            done
+            echo "$map predictions for $hemisphere are done!"
         done
-    done
+    fi
 done
+echo "All done!"
